+++ src/server/index.mjs
@@ -11,6 +11,8 @@ import Context from "@prisma-cms/prisma-context";
 
 import fs from "fs";
 
+import chalk from "chalk";
+
 // import Knex from "knex";
 
 import { ImagesMiddleware } from "@prisma-cms/upload-module";
@@ -29,7 +31,74 @@ const resolvers = cmsModule.getResolvers();
 const endpoint = process.env.endpoint;
 
 
-export const modifyArgs = function (where, modifier, info) {
+export const modifyArgs = function (source, args, ctx, info, modifier) {
+
+  if (arguments.length <= 3) {
+
+    return modifyArgsDeprecated.apply(modifyArgsDeprecated, arguments);
+
+  }
+
+  else {
+
+    let where = args.where;
+
+    if (where && modifier) {
+
+      /**
+       * Очищаем все аргументы, иначе вычищаемый параметр будет взят в дальнейшем
+       * из объекта info.
+       * Делать это надо только один раз
+       */
+
+      if (info) {
+
+        info.fieldNodes.map(n => {
+          n.arguments = []
+        });
+      }
+
+      return modifyArgsNew(source, args, ctx, info, modifier, where);
+
+    }
+
+  }
+
+}
+
+
+const modifyArgsNew = function (source, args, ctx, info, modifier, where) {
+
+
+  if (typeof where === "object") {
+
+    if (Array.isArray(where)) {
+
+      where.map(n => modifyArgsNew(source, args, ctx, info, modifier, n))
+
+    }
+    else {
+
+      modifier(source, args, ctx, info, where);
+
+      /**
+       * Проходим по всем остальным элементам условия
+       */
+      for (var i in where) {
+        modifyArgsNew(source, args, ctx, info, modifier, where[i]);
+      }
+
+    }
+
+  }
+
+
+  return where;
+}
+
+
+const modifyArgsDeprecated = function (where, modifier, info) {
+
 
   if (where && modifier) {
 
@@ -41,6 +110,9 @@ export const modifyArgs = function (where, modifier, info) {
      */
 
     if (info) {
+
+      console.error(chalk.red("Deprecated"), "modifyArgs with 3 or less arguments is deprecated. Use `source, args, ctx, info` instead.");
+
       info.fieldNodes.map(n => {
         n.arguments = []
       });
@@ -50,7 +122,7 @@ export const modifyArgs = function (where, modifier, info) {
 
       if (Array.isArray(where)) {
 
-        where.map(n => modifyArgs(n, modifier))
+        where.map(n => modifyArgsDeprecated(n, modifier))
 
       }
       else {
@@ -61,7 +133,7 @@ export const modifyArgs = function (where, modifier, info) {
          * Проходим по всем остальным элементам условия
          */
         for (var i in where) {
-          modifyArgs(where[i], modifier);
+          modifyArgsDeprecated(where[i], modifier);
         }
 
       }
@@ -201,7 +273,7 @@ export class PrismaCmsServer {
         resolverValidationOptions: {
           requireResolversForResolveType: false,
         },
-        context,
+        context: request => this.processRequest(request),
         resolvers,
         ...serverOptions,
       });
@@ -214,6 +286,12 @@ export class PrismaCmsServer {
     }
 
     return this.server;
+  }
+
+
+  processRequest(request) {
+
+    return this.context(request);
   }