+++ changelog.md
@@ -1,3 +1,8 @@
+2.3.1
+=====================================
+- Fix request context
+- modifyArgs refactoring
+
 2.3.0
 =====================================
 - Added modifyArgs()